+++ common/src/main/java/tocraft/craftedcore/event/common/CommandEvents.java
@@ -2,7 +2,7 @@ package tocraft.craftedcore.event.common;
 
 import com.mojang.brigadier.CommandDispatcher;
 //#if MC>1182
-//$$ import net.minecraft.commands.CommandBuildContext;
+import net.minecraft.commands.CommandBuildContext;
 //#endif
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
@@ -16,9 +16,9 @@ public final class CommandEvents {
     @FunctionalInterface
     public interface CommandRegistration {
         //#if MC>1182
-        //$$ void register(CommandDispatcher<CommandSourceStack> dispatcher, CommandBuildContext registry, Commands.CommandSelection selection);
+        void register(CommandDispatcher<CommandSourceStack> dispatcher, CommandBuildContext registry, Commands.CommandSelection selection);
         //#else
-        void register(CommandDispatcher<CommandSourceStack> dispatcher, Commands.CommandSelection selection);
+        //$$ void register(CommandDispatcher<CommandSourceStack> dispatcher, Commands.CommandSelection selection);
         //#endif
     }
 }