+++ common/src/main/java/tocraft/craftedcore/mixin/client/ClientPacketListenerMixin.java
@@ -4,9 +4,9 @@ import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.Minecraft;
 //#if MC>1201
-//$$ import net.minecraft.client.multiplayer.ClientCommonPacketListenerImpl;
-//$$ import net.minecraft.client.multiplayer.CommonListenerCookie;
-//$$ import net.minecraft.network.Connection;
+import net.minecraft.client.multiplayer.ClientCommonPacketListenerImpl;
+import net.minecraft.client.multiplayer.CommonListenerCookie;
+import net.minecraft.network.Connection;
 //#endif
 import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.minecraft.client.player.LocalPlayer;
@@ -25,21 +25,21 @@ import tocraft.craftedcore.event.client.ClientPlayerEvents;
 @Environment(EnvType.CLIENT)
 @Mixin(ClientPacketListener.class)
 //#if MC>1201
-//$$ public abstract class ClientPacketListenerMixin extends ClientCommonPacketListenerImpl {
-//$$     protected ClientPacketListenerMixin(Minecraft minecraft, Connection connection, CommonListenerCookie commonListenerCookie) {
-//$$         super(minecraft, connection, commonListenerCookie);
-//$$     }
+public abstract class ClientPacketListenerMixin extends ClientCommonPacketListenerImpl {
+    protected ClientPacketListenerMixin(Minecraft minecraft, Connection connection, CommonListenerCookie commonListenerCookie) {
+        super(minecraft, connection, commonListenerCookie);
+    }
 //#else
-public abstract class ClientPacketListenerMixin {
+//$$ public abstract class ClientPacketListenerMixin {
 //#endif
 
     @Unique
     private LocalPlayer craftedcore$oldPlayer = null;
 
     //#if MC<=1201
-    @Shadow
-    @Final
-    private Minecraft minecraft;
+    //$$ @Shadow
+    //$$ @Final
+    //$$ private Minecraft minecraft;
     //#endif
 
     @Inject(method = "handleLogin", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/Options;setServerRenderDistance(I)V", shift = At.Shift.AFTER))
@@ -54,9 +54,9 @@ public abstract class ClientPacketListenerMixin {
 
     @Inject(method = "handleRespawn", at = @At(value = "INVOKE",
             //#if MC>1201
-            //$$ target = "Lnet/minecraft/client/multiplayer/ClientLevel;addEntity(Lnet/minecraft/world/entity/Entity;)V"))
+            target = "Lnet/minecraft/client/multiplayer/ClientLevel;addEntity(Lnet/minecraft/world/entity/Entity;)V"))
             //#else
-            target = "Lnet/minecraft/client/multiplayer/ClientLevel;addPlayer(ILnet/minecraft/client/player/AbstractClientPlayer;)V"))
+            //$$ target = "Lnet/minecraft/client/multiplayer/ClientLevel;addPlayer(ILnet/minecraft/client/player/AbstractClientPlayer;)V"))
     //#endif
     private void handleRespawn(ClientboundRespawnPacket packet, CallbackInfo ci) {
         ClientPlayerEvents.CLIENT_PLAYER_RESPAWN.invoke().respawn(this.craftedcore$oldPlayer, minecraft.player);