+++ common/src/main/java/tocraft/craftedcore/network/ModernNetworking.java
@@ -8,15 +8,18 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.minecraft.nbt.CompoundTag;
 //#if MC>=1205
-//$$ import net.minecraft.network.RegistryFriendlyByteBuf;
-//$$ import net.minecraft.network.codec.StreamCodec;
+import net.minecraft.network.RegistryFriendlyByteBuf;
+import net.minecraft.network.codec.StreamCodec;
 //#endif
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.protocol.Packet;
 //#if MC>1201
-//$$ import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
-//$$ import java.util.Map;
-//$$ import java.util.HashMap;
+import net.minecraft.network.protocol.common.ClientboundCustomPayloadPacket;
+import net.minecraft.network.protocol.common.ServerboundCustomPayloadPacket;
+import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
+import java.util.Map;
+import java.util.HashMap;
+import org.jetbrains.annotations.NotNull;
 //#endif
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
@@ -26,7 +29,7 @@ import org.jetbrains.annotations.ApiStatus;
 @SuppressWarnings("unused")
 public class ModernNetworking {
     //#if MC>=1205
-    //$$ private static final Map<ResourceLocation, CustomPacketPayload.Type<PacketPayload>> TYPES = new HashMap<>();
+    private static final Map<ResourceLocation, CustomPacketPayload.Type<PacketPayload>> TYPES = new HashMap<>();
     //#endif
 
     @ExpectPlatform
@@ -35,30 +38,26 @@ public class ModernNetworking {
     }
 
     //#if MC>=1205
-    //$$ @ExpectPlatform
-    //$$ public static void registerType(ResourceLocation id) {
-    //$$     throw new AssertionError();
-    //$$ }
-    //$$
-    //$$ public static CustomPacketPayload.Type<PacketPayload> getType(ResourceLocation id) {
-    //$$     if (!TYPES.containsKey(id)) {
-    //$$         TYPES.put(id, new CustomPacketPayload.Type<>(id));
-    //$$     }
-    //$$     return TYPES.get(id);
-    //$$ }
+    @ExpectPlatform
+    public static void registerType(ResourceLocation id) {
+        throw new AssertionError();
+    }
+    
+    public static CustomPacketPayload.Type<PacketPayload> getType(ResourceLocation id) {
+        if (!TYPES.containsKey(id)) {
+            TYPES.put(id, new CustomPacketPayload.Type<>(id));
+        }
+        return TYPES.get(id);
+    }
     //#endif
 
     public static void sendToPlayer(ServerPlayer player, ResourceLocation packetId, CompoundTag data) {
-        //#if MC>1201
-        //$$ FriendlyByteBuf buf = new FriendlyByteBuf(Unpooled.buffer());
-        //$$ buf.writeResourceLocation(packetId);
-        //$$ buf.writeNbt(data);
-        //$$ player.connection.send(toPacket(Side.S2C, packetId, buf));
+        //#if MC>=1205
+        player.connection.send(toPacket(Side.S2C, new PacketPayload(packetId, data)));
+        //#elseif MC>1201
+        //$$ player.connection.send(toPacket(Side.S2C, packetId, data));
         //#else
-        FriendlyByteBuf buf = new FriendlyByteBuf(Unpooled.buffer());
-        buf.writeResourceLocation(packetId);
-        buf.writeNbt(data.copy());
-        player.connection.send(toPacket(Side.S2C, packetId, buf));
+        //$$ player.connection.send(toPacket(Side.S2C, packetId, data));
         //#endif
     }
 
@@ -73,25 +72,29 @@ public class ModernNetworking {
         ClientPacketListener connection = Minecraft.getInstance().getConnection();
 
         if (connection != null) {
-            //#if MC>1201
-            //$$ FriendlyByteBuf buf = new FriendlyByteBuf(Unpooled.buffer());
-            //$$ buf.writeResourceLocation(packetId);
-            //$$ buf.writeNbt(data);
-            //$$ connection.send(toPacket(Side.C2S, packetId, buf));
+            //#if MC>=1205
+            connection.send(toPacket(Side.C2S, new PacketPayload(packetId, data)));
+            //#elseif MC>1201
+            //$$ connection.send(toPacket(Side.C2S, packetId, data));
             //#else
-            FriendlyByteBuf buf = new FriendlyByteBuf(Unpooled.buffer());
-            buf.writeResourceLocation(packetId);
-            buf.writeNbt(data.copy());
-            connection.send(toPacket(Side.C2S, packetId, buf));
+            //$$ connection.send(toPacket(Side.C2S, packetId, data));
             //#endif
         }
     }
 
+    //#if MC>=1205
     @ExpectPlatform
     @ApiStatus.Internal
-    public static Packet<?> toPacket(ModernNetworking.Side side, ResourceLocation id, FriendlyByteBuf buf) {
+    public static Packet<?> toPacket(ModernNetworking.Side side, CustomPacketPayload payload) {
         throw new AssertionError();
     }
+    //#else
+    //$$ @ExpectPlatform
+    //$$ @ApiStatus.Internal
+    //$$ public static Packet<?> toPacket(ModernNetworking.Side side, ResourceLocation id, CompoundTag data) {
+    //$$     throw new AssertionError();
+    //$$ }
+    //#endif
 
     @FunctionalInterface
     public interface Receiver {
@@ -115,11 +118,11 @@ public class ModernNetworking {
     }
 
     //#if MC>1201
-    //$$ @ApiStatus.Internal
-    //$$ public record PacketPayload(ResourceLocation id,
-    //$$                             CompoundTag nbt) implements CustomPacketPayload {
+    @ApiStatus.Internal
+    public record PacketPayload(ResourceLocation id,
+                                CompoundTag nbt) implements CustomPacketPayload {
         //#if MC>=1205
-        //$$ public void write(RegistryFriendlyByteBuf buf) {
+        public void write(RegistryFriendlyByteBuf buf) {
         //#else
         //$$ public PacketPayload(FriendlyByteBuf buf) {
         //$$     this(buf.readResourceLocation(), buf.readNbt());
@@ -127,35 +130,35 @@ public class ModernNetworking {
         //$$
         //$$ public void write(FriendlyByteBuf buf) {
             //#endif
-    //$$         buf.writeResourceLocation(id);
-    //$$         buf.writeNbt(nbt);
-    //$$     }
-    //$$
+            buf.writeResourceLocation(id);
+            buf.writeNbt(nbt);
+        }
+    
         //#if MC>=1205
-        //$$
-        //$$ public PacketPayload(RegistryFriendlyByteBuf buf) {
-        //$$     this(buf.readResourceLocation(), buf.readNbt());
-        //$$ }
-        //$$ @Override
-        //$$ public @NotNull Type<? extends CustomPacketPayload> type() {
-        //$$     return getType(id);
-        //$$ }
-        //$$
-        //$$ public static StreamCodec<RegistryFriendlyByteBuf, PacketPayload> streamCodec() {
-        //$$     return new StreamCodec<>() {
-        //$$         @Override
-        //$$         public @NotNull PacketPayload decode(@NotNull RegistryFriendlyByteBuf buf) {
-        //$$             return new PacketPayload(buf);
-        //$$         }
-        //$$
-        //$$         @Override
-        //$$         public void encode(@NotNull RegistryFriendlyByteBuf buf, @NotNull PacketPayload payload) {
-        //$$             buf.writeResourceLocation(payload.id);
-        //$$             buf.writeNbt(payload.nbt);
-        //$$         }
-        //$$     };
-        //$$ }
+        
+        public PacketPayload(RegistryFriendlyByteBuf buf) {
+            this(buf.readResourceLocation(), buf.readNbt());
+        }
+        @Override
+        public @NotNull Type<? extends CustomPacketPayload> type() {
+            return getType(id);
+        }
+        
+        public static StreamCodec<RegistryFriendlyByteBuf, PacketPayload> streamCodec() {
+            return new StreamCodec<>() {
+                @Override
+                public @NotNull PacketPayload decode(@NotNull RegistryFriendlyByteBuf buf) {
+                    return new PacketPayload(buf);
+                }
+        
+                @Override
+                public void encode(@NotNull RegistryFriendlyByteBuf buf, @NotNull PacketPayload payload) {
+                    buf.writeResourceLocation(payload.id);
+                    buf.writeNbt(payload.nbt);
+                }
+            };
+        }
         //#endif
-    //$$ }
+    }
     //#endif
 }