+++ common/src/main/java/tocraft/craftedcore/patched/CCommandSourceStack.java
@@ -1,0 +1,15 @@
+package tocraft.craftedcore.patched;
+
+import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.network.chat.Component;
+
+@SuppressWarnings("unused")
+public class CCommandSourceStack {
+    public static void sendSuccess(CommandSourceStack source, Component text, boolean bl) {
+        //#if MC>=1201
+        source.sendSuccess(() -> text, bl);
+        //#else
+        //$$ source.sendSuccess(text, bl);
+        //#endif
+    }
+}