+++ common/src/main/java/tocraft/craftedcore/patched/CEntity.java
@@ -6,17 +6,17 @@ import net.minecraft.world.level.Level;
 public class CEntity {
     public static Level level(Entity entity) {
         //#if MC>1194
-        //$$ return entity.level();
+        return entity.level();
         //#else
-        return entity.level;
+        //$$ return entity.level;
         //#endif
     }
 
     public static boolean isOnGround(Entity entity) {
         //#if MC>1194
-        //$$ return entity.onGround();
+        return entity.onGround();
         //#else
-        return entity.isOnGround();
+        //$$ return entity.isOnGround();
         //#endif
     }
 }