+++ common/src/main/java/tocraft/craftedcore/registration/SynchronizedReloadListenerRegistry.java
@@ -33,8 +33,8 @@ public class SynchronizedReloadListenerRegistry {
 
 
         //#if MC>=1205
-        //$$ // register Network packet
-        //$$ ModernNetworking.registerType(reloadListener.RELOAD_SYNC);
+        // register Network packet
+        ModernNetworking.registerType(reloadListener.RELOAD_SYNC);
         //#endif
 
         onRegister(reloadListener, id);