+++ fabric/src/main/java/tocraft/craftedcore/fabric/mixin/client/MinecraftClientMixin.java
@@ -15,9 +15,9 @@ import tocraft.craftedcore.event.client.ClientPlayerEvents;
 @Mixin(Minecraft.class)
 public abstract class MinecraftClientMixin {
     //#if MC>=1205
-    //$$ @Inject(method = "disconnect(Lnet/minecraft/client/gui/screens/Screen;Z)V",
-    //$$                 at = @At(value = "INVOKE", target = "Lnet/minecraft/client/GameNarrator;clear()V"))
-    //$$ private void onDisconnect(Screen screen, boolean retainDownloadedPacks, CallbackInfo ci) {
+    @Inject(method = "disconnect(Lnet/minecraft/client/gui/screens/Screen;Z)V",
+                    at = @At(value = "INVOKE", target = "Lnet/minecraft/client/GameNarrator;clear()V"))
+    private void onDisconnect(Screen screen, boolean retainDownloadedPacks, CallbackInfo ci) {
     //#elseif MC>1201
     //$$     @Inject(method = "disconnect(Lnet/minecraft/client/gui/screens/Screen;)V",
     //$$             at = @At(value = "INVOKE", target = "Lnet/minecraft/client/GameNarrator;clear()V"))
@@ -27,9 +27,9 @@ public abstract class MinecraftClientMixin {
     //$$                 at = @At(value = "INVOKE", target = "Lnet/minecraft/client/GameNarrator;clear()V"))
     //$$         private void onDisconnect(Screen nextScreen, CallbackInfo ci) {
     //#else
-    @Inject(method = "clearLevel(Lnet/minecraft/client/gui/screens/Screen;)V",
-            at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/chat/NarratorChatListener;clear()V"))
-    private void onDisconnect(Screen nextScreen, CallbackInfo ci) {
+    //$$ @Inject(method = "clearLevel(Lnet/minecraft/client/gui/screens/Screen;)V",
+    //$$         at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/chat/NarratorChatListener;clear()V"))
+    //$$ private void onDisconnect(Screen nextScreen, CallbackInfo ci) {
     //#endif
         ClientPlayerEvents.CLIENT_PLAYER_QUIT.invoke().quit(Minecraft.getInstance().player);
     }