+++ forge/src/main/java/tocraft/craftedcore/forge/client/CraftedCoreForgeEventHandlerClient.java
@@ -6,17 +6,17 @@ import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.client.event.ClientPlayerNetworkEvent;
 //#if MC>1194
-//$$ import net.minecraftforge.client.event.CustomizeGuiOverlayEvent;
+import net.minecraftforge.client.event.CustomizeGuiOverlayEvent;
 //#elseif MC>1182
 //$$ import net.minecraftforge.client.event.RenderGuiEvent;
 //#else
-import net.minecraftforge.client.event.RenderGameOverlayEvent;
+//$$ import net.minecraftforge.client.event.RenderGameOverlayEvent;
 //#endif
 //#if MC>1182
-//$$ import net.minecraftforge.client.event.RenderGuiOverlayEvent;
+import net.minecraftforge.client.event.RenderGuiOverlayEvent;
 //#else
-import net.minecraftforge.client.gui.OverlayRegistry;
-import com.mojang.blaze3d.vertex.PoseStack;
+//$$ import net.minecraftforge.client.gui.OverlayRegistry;
+//$$ import com.mojang.blaze3d.vertex.PoseStack;
 //#endif
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -29,17 +29,17 @@ import tocraft.craftedcore.event.client.RenderEvents;
 public class CraftedCoreForgeEventHandlerClient {
     @SubscribeEvent
     //#if MC>1194
-    //$$ public void event(CustomizeGuiOverlayEvent event) {
-    //$$     RenderEvents.HUD_RENDERING.invoke().render(event.getGuiGraphics(), event.getPartialTick());
-    //$$ }
+    public void event(CustomizeGuiOverlayEvent event) {
+        RenderEvents.HUD_RENDERING.invoke().render(event.getGuiGraphics(), event.getPartialTick());
+    }
     //#elseif MC>1182
     //$$ public void event(RenderGuiEvent.Post event) {
     //$$     RenderEvents.HUD_RENDERING.invoke().render(event.getPoseStack(), event.getPartialTick());
     //$$ }
     //#else
-    public void event(RenderGameOverlayEvent.Post event) {
-        RenderEvents.HUD_RENDERING.invoke().render(event.getMatrixStack(), event.getPartialTicks());
-    }
+    //$$ public void event(RenderGameOverlayEvent.Post event) {
+    //$$     RenderEvents.HUD_RENDERING.invoke().render(event.getMatrixStack(), event.getPartialTicks());
+    //$$ }
     //#endif
 
     @SubscribeEvent
@@ -53,25 +53,25 @@ public class CraftedCoreForgeEventHandlerClient {
 
     @SubscribeEvent
     //#if MC>1182
-    //$$ public void event(RenderGuiOverlayEvent.Pre event) {
+    public void event(RenderGuiOverlayEvent.Pre event) {
     //#else
-    public void event(RenderGameOverlayEvent.PreLayer event) {
+    //$$ public void event(RenderGameOverlayEvent.PreLayer event) {
     //#endif
         //#if MC>1194
-        //$$ var graphics = event.getGuiGraphics();
+        var graphics = event.getGuiGraphics();
         //#elseif MC>1182
         //$$ var graphics = event.getPoseStack();
         //#else
-        var graphics = event.getMatrixStack();
+        //$$ var graphics = event.getMatrixStack();
         //#endif
         //#if MC>1182
-        //$$ switch (event.getOverlay().id().getPath()) {
-        //$$     case "player_health" -> {
+        switch (event.getOverlay().id().getPath()) {
+            case "player_health" -> {
         //#else
-        OverlayRegistry.OverlayEntry overlayEntry = OverlayRegistry.getEntry(event.getOverlay());
-        if (overlayEntry != null)
-        switch (overlayEntry.getDisplayName()) {
-            case "Player Health" -> {
+        //$$ OverlayRegistry.OverlayEntry overlayEntry = OverlayRegistry.getEntry(event.getOverlay());
+        //$$ if (overlayEntry != null)
+        //$$ switch (overlayEntry.getDisplayName()) {
+        //$$     case "Player Health" -> {
         //#endif
                 InteractionResult result = RenderEvents.RENDER_HEALTH.invoke().render(graphics, Minecraft.getInstance().player);
                 if (result == InteractionResult.FAIL) {
@@ -79,9 +79,9 @@ public class CraftedCoreForgeEventHandlerClient {
                 }
             }
             //#if MC>1182
-            //$$ case "food_level" -> {
+            case "food_level" -> {
             //#else
-            case "Food Level" -> {
+            //$$ case "Food Level" -> {
             //#endif
                 InteractionResult result = RenderEvents.RENDER_FOOD.invoke().render(graphics, Minecraft.getInstance().player);
                 if (result == InteractionResult.FAIL) {
@@ -89,9 +89,9 @@ public class CraftedCoreForgeEventHandlerClient {
                 }
             }
             //#if MC>1182
-            //$$ case "air_level" -> {
+            case "air_level" -> {
             //#else
-            case "Air Level" -> {
+            //$$ case "Air Level" -> {
                 //#endif
                 InteractionResult result = RenderEvents.RENDER_BREATH.invoke().render(graphics, Minecraft.getInstance().player);
                 if (result == InteractionResult.FAIL) {
@@ -99,9 +99,9 @@ public class CraftedCoreForgeEventHandlerClient {
                 }
             }
         //#if MC>1182
-        //$$ case "vehicle_health" -> {
+        case "vehicle_health" -> {
             //#else
-            case "Mount Health" -> {
+            //$$ case "Mount Health" -> {
                 //#endif
                 InteractionResult result = RenderEvents.RENDER_MOUNT_HEALTH.invoke().render(graphics, Minecraft.getInstance().player);
                 if (result == InteractionResult.FAIL) {
@@ -112,9 +112,9 @@ public class CraftedCoreForgeEventHandlerClient {
     }
     @SubscribeEvent
     //#if MC>1182
-    //$$ public void event(ClientPlayerNetworkEvent.LoggingOut event) {
+    public void event(ClientPlayerNetworkEvent.LoggingOut event) {
     //#else
-    public void event(ClientPlayerNetworkEvent.LoggedOutEvent event) {
+    //$$ public void event(ClientPlayerNetworkEvent.LoggedOutEvent event) {
     //#endif
         ClientPlayerEvents.CLIENT_PLAYER_QUIT.invoke().quit(event.getPlayer());
     }